Defer typing + move generateSQL to a Worker
Filter typing freeze:
- useDeferredValue on matchedTrades + recentMatchedTrades inside
  SeriesStatsTable. React renders the OLD seriesData immediately on
  keystroke, then schedules the recompute at low priority. Typing
  stays smooth even when each commit takes >100ms.

Generate SQL freeze:
- Extract the 340-line generator into src/utils/generateSqlBody.ts
  as a pure function.
- New src/workers/generateSql.worker.ts runs it off the main thread.
- Click handler dispatches to the worker, shows "Generating..." on
  the button, sets the modal when the worker posts back.
- Drop the legacy inline implementation entirely.

No behavior change in the SQL output — same logic, just no longer
freezing the page for the 5-10s of backtest passes.
